Coverage: Edge-graph coverage optional functionality for UTM conversion.
The combination of optional simplified input graph alongside the UTM conversion is a necessity for splitmerge:
* splitmerge must act on simplified edges (of which the original simplified is cut into a few subcurves).
* Partial curve matching (which is part of the edge-wise coverage computation) works better when curves are represented in UTM coordinates, yet I haven't build the logic to apply that transformation on a simplified graph, so it requires a separate option.

Note how both simplification and UTM conversion can be derived from the input graph. But for sanity checking purposes I enforce to provide this as a parameter to prevent errors down the line when working with accidentally incorrectly formatted graphs in workflows.