APLS: Filter out precomputed distance matrix to contain only the relevant control nodes.
Start-end node pairs `u-v` should always have `v > u` and both `u` and `v` contained in the control nodes set.
Having this filter applied in the construction of the distance matrix prevents the need to perform this filter multiple times later on in the code where the distance matrix is read from.