APLS: Bugfix: Reducing the node identifiers of `G_to_H` to the set of control points.
When constructing samples for the three categories (A, B, C), the code iterates all shortest path entries it has precomputed, to ensure we only take samples which are either reachable or unreachable.
Due to computation intensity, shortest path computations is only performed on the selection of control nodes.
As we iterate the `G_to_H` dictionary keys, it is therefore revelant to reduce this set to the control nodes, to prevent looking up the existence of a shortest path between node identifiers we have not attempted to compute the shortest path for.