Deduplication: Bugfix: Incorrect edge removal and injection causing nodes without position.
When we delete a selection of nodes from a graph, both the dangling edges (both endpoints are connected to a node to be deleted) and partially dangling edges (one endpoint is connected to an node to be deleted) will be deleted from the graph.

In the logic first such (partially) dangling edges were deleted and then updated (partially) dangling edges were reintroduced. The logic of NetworkX however just inserts edges even if their endpoints do not exist in the graph, if that is the case a node without attributes (thus neither x and y position) is injected.

The fix is to
1. Iterate all (partially) dangling edges, and replace the deprecated nid with the unique nid.
2. Inject these new edges.
3. Remove the duplicated nodes from the graph.