APLS: Bugfix: Control nodes used for `Hc` were not filtering out nids of `G` pointing at `None`.
The distance matrix is computed on all control nodes.
This is first done on `G` without issue, yet then we map _all_ control nodes of `G` to `Hc` without any filtering.
This likely causes an error, because if no relation exists between a control node of `G` to `Hc`, the distance matrix is attempted to compute on node `None`.

Solved by simply filtering out all `G_to_Hc[G_nid] != None` to conform to valid and complete set of control nodes of `Hc`.