Node position extraction: Refactor: Extraction as dictionary  (by nid) or list more concise naming to prevent confusion in picking incorrect function.
Some of my logic got bugs, because it would use the node positions data like a dictionary, while it was provided as a list.
The program doesn't crash, because the nid identifier can be used as an list index. However, the node position returned therefore (most likely) is incorrect.

Solved by explicitly appending "dictionary" or "list" to the "extract_node_positions" functions.