Simplification: Optionally retain edge attributes.
In the simplification process we can just drop all edge attributes (and then afterwards recompute the curvature, geometry and length).

However, sometimes an intermediate graph should stick to its attributes.
There is a sanity check that whatever of such attribute is found, that the attribute value amongst all edges (if they have it) is consistent.

At this moment I think that if we would be merging edges with different attribute values is a code smell.