Vectorization: Track and annotate every edge with the "vectorized_from" attribute.
This is necessary for the reconnection step in graph merging.
Namely, we only have to reconnect at nodes which have at least two connected edges deleted due to the same injected edge.
This requires that we know what edge it was covered by, which requires that the vectorized graph (within the edge coverage computation) knows where its coming from.