Merging: Performance edge endpoint reconnecting: Re-use node and edge tree.
When injecting a node, extend the node tree with this new node and extend the edge tree with the two new edges. Additionally add the removed edge to the set of edges to ignore.

This allows to re-use the trees which have sufficient performance improvement to use.

More logical would be to remove the edge from the edge tree, but for some magical reason `rtree` function then crashes...

Furthermore simplify `format_eid` logic.

* Re-use node and edge tree by reconnecting nodes.
* Simplify `format_eid` logic.