Merging: Duplicate removal: Don't delete edges covered by connection points.
* Exclude from edges for removal those which are covered by connection nodes (B to C).
* New function `edges_covered_by_nid` to find edges covered by a specified node.

E.g. small alleys got their edges deleted, because their entire curvature was within range of the connection point. Simply filtering out those edges significantly improve this issue.

A remaining problem is that adjacent/parallel curvature is removed by an injected edge which is not near the connection point (thus not protected from deletion). This could be solved by checking that the edges appointed for removal are connected to the same path. Yet this would introduce new issues and this in itself should get tackled by node reconnection.