Graph curvature: Refactor + Debug.
* Converting `assert` into `check` statements.
* Ensure injected `eids` don't have a key alongside.

There is a case where we cut the curvature of a self-looping edge in half.
This results in two edges with `(u, w)` and `(w, u)`.
When we apply `format_eid`, both edges will be converted into `(u, w, 0)`, and since their key is presen the `G.add_edges_from` will overwrite the first injection with the second, causing a loss of an edge (and its curvature) of the graph.

Fixed by sorting the edge endpoint node identifiers without adding a key. The injection logic of NetworkX automatically increments the key value of subsequent occurrences of the same `(u, v)` pair.