Measurements: Fix: Merge GPS into Sat rather than against the Sat-vs-GPS pre-filtered graph.
The merging procedures are applied to a selection of sat edges where gps edges occur. By prefiltering the sat graph it is easier to focus on what the behaviour/effect of these merging functions are. However, the actual goal is to improve the sat graph without filtering out subgraphs where no gps edges live. So this prefilter is just for debugging and not the actual usage goal.

Fixed it by adding the ability to prefilter for debugging purposes, but by default don't prefilter (thus inject gps edges onto the entire sat graph).